setup train

+++ openrec/postprocess/ctc_postprocess.py
@@ -104,7 +104,7 @@ class CTCLabelDecode(BaseRecLabelDecode):
     def __call__(self, preds, batch=None, **kwargs):
         # preds = preds['res']
         if kwargs.get('torch_tensor', True):
-            preds = preds.detach().cpu().numpy()
+            preds = preds.detach().cpu().float().numpy()
         preds_idx = preds.argmax(axis=2)
         preds_prob = preds.max(axis=2)
         text = self.decode(preds_idx, preds_prob, is_remove_duplicate=True)
